+++ src/index.js
@@ -1,1 +1,3 @@
 export { styleguidist } from './styleguidist.js';
+export { start } from './start.js';
+export { takeStyleguidistSnapshots } from './snapshots.js';